fix(orders): ship orders/order-requests punch-list (C2-C4, C7-C8, I1-I15, M1/M3/M6/M7/M8)
Service layer (apps/web/src/server/services/order-requests.ts):
- C2: cancel() now gates on assertPermission('orders:request')
- C3: approve/deny/setStatus/markDelivered/setInternalNotes now go
  through new private requireWarehouseAccess() helper, which loads
  the request's warehouse_id and asserts read/write access in the
  caller's scope before any mutation
- I2: documented the public-link no-cancel-path convention inline
- I4: replaced misleading "Edit line qtys" error in markDelivered
  with "Adjust stock first, then retry."
- M1: added secondary `id desc` sort to list() for stable pagination
- M3: capped offset at 10_000 in list()
- M6: parallelized lines+reservations+warehouse reads in get()
- M7: requesters can only self-cancel a request that is still
  pending_approval — managers unaffected
- M8: setStatus('ready_for_delivery') now bails if the destination
  warehouse has been archived

Server actions (apps/web/src/server/actions/order-requests.ts):
- I14: createSchema.lines.quantity now uses .int() for integer-only

Public POST route (apps/web/src/app/api/v1/public/order-requests/route.ts):
- C7: only trust x-forwarded-for when VERCEL === '1' — outside Vercel
  fall back to x-real-ip or 'unknown'
- C8: lowercase requester_email on insert; reuse the canonical
  address in the confirmation email + trackUrl
- I5: dedupe lines BEFORE the total-qty cap so duplicate-row
  spammers can't trip the cap with collapsible payload
- I13: hidden honeypot field (hp) — non-empty values return a
  silent 200 so bots can't distinguish success from failure
- I14: lineSchema.quantity now .int()

Public GET track route (apps/web/src/app/api/v1/public/order-requests/[id]/route.ts):
- I9: replaced raw manager-typed denied_reason on the public track
  payload with a generic "Your request was not approved." Reason
  text is now manager-only; share-with-requester opt-in flag is a
  future enhancement.

Components:
- I8: CancelOrderButton now includes an optional reason textarea
  inside the destructive-confirm dialog and forwards the trimmed
  value to cancelOrderRequestAction
- I13: PublicOrderForm renders an off-screen honeypot input and
  sends its value as `hp` in the POST body
- I15: PublicOrderForm no longer double-appends `&t=` — the server
  already folds the token into trackUrl
- I11: PublicTokenControls renders a 90-day rotation reminder
  banner when public_request_token_rotated_at is older than 90
  days; banner clears immediately on regenerate

Migrations:
- 0076: BEFORE UPDATE trigger _validate_order_request_status_transition
  rejects non-whitelisted status transitions (closes C4 — RLS UPDATE
  policy hardening from 0045 lacked transition guard)
- 0077: cancel_order_request reverts to clear-or-replace semantics
  for denied_reason (I1 — 0055 regressed the 0045 L2 fix)
- 0078: tightened order_request_lines INSERT RLS to require the
  inserted item.warehouse_id = parent request.warehouse_id (I7)

Skipped per brief / out of scope:
- C1: handled by Shipments agent in 0073
- I6: lock ordering already standardized to `order by l.item_id`
  in 0044/0045/0055
- I16: notifications RLS already correctly scopes to user_id =
  auth.uid() per 0003_rls.sql lines 221-224
- M10: shipment_lines.qty_shipped sum belongs to Shipments territory
- M2/M4/M12: explicitly optional in brief

+++ apps/web/src/server/actions/order-requests.ts
@@ -27,7 +27,12 @@ const createSchema = z
       .array(
         z.object({
           itemId: z.string().uuid(),
-          quantity: z.coerce.number().positive().max(MAX_QTY_PER_LINE),
+          // I14: every order line is an integer count of books — fractional
+          // requests like 1.5 don't represent anything sensible and would
+          // confuse downstream stock-movement math. `.int()` rejects them
+          // outright; `.coerce` keeps tolerating string inputs from form
+          // posts.
+          quantity: z.coerce.number().int().positive().max(MAX_QTY_PER_LINE),
           notes: z.string().max(500).nullable().optional(),
         }),
       )
